third_party/WebKit/LayoutTests/scrollbars/custom-scrollbar-not-inherited-by-iframe-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS 285 is 285
